rest

+++ src/backend/services/admin/DeveloperService.ts
@@ -1,19 +1,13 @@
-import {IDeveloper} from "../../db/models/DeveloperModel";
 import {IDeveloperStorageService} from "../../db/storageServices/DeveloperStorageService";
 import {ISubscribeStorageService} from "../../db/storageServices/SubscribeStorageService";
-
-
-export interface IListQuery {
-    search?: string;
-    limit?: number;
-    channelId: string;
-}
+import {IFavoriteDeveloper, INewFavoriteDeveloper} from "../../db/models/developer/favorite/FavoriteDeveloperModel";
+import {IListQuery} from "./IListQuery";
 
 
 export interface IDeveloperService {
-    list(query: IListQuery): Promise<IDeveloper[]>;
-    add(obj: IDeveloper): Promise<boolean>;
-    delete(obj: {channelId: string, username: string}): Promise<boolean>;
+    list(query: IListQuery<IFavoriteDeveloper>): Promise<IFavoriteDeveloper[]>;
+    add(obj: INewFavoriteDeveloper): Promise<boolean>;
+    delete(obj: Partial<IFavoriteDeveloper>): Promise<boolean>;
 }
 
 
@@ -26,12 +20,12 @@ export default class DeveloperService implements IDeveloperService {
         this.subscribeStorageService = subscribeStorageService;
     }
 
-    async list(query: IListQuery): Promise<IDeveloper[]> {
-        const {channelId, search, limit} = query;
-        return await this.developerStorageService.get({channelId}, search, limit);
+    async list(query: IListQuery<IFavoriteDeveloper>): Promise<IFavoriteDeveloper[]> {
+        const {search, limit, filter} = query;
+        return await this.developerStorageService.get(filter, search, limit);
     }
 
-    async add(obj: IDeveloper): Promise<boolean> {
+    async add(obj: INewFavoriteDeveloper): Promise<boolean> {
         if (obj.username.length !== 0) {
             return this.developerStorageService.add(obj);
         } else {
@@ -39,9 +33,13 @@ export default class DeveloperService implements IDeveloperService {
         }
     }
 
-    async delete(obj: {channelId: string, username: string}): Promise<boolean> {
-        const {username, channelId} = obj;
-        const developerRemovingOperationResult = await this.developerStorageService.remove(obj);
+    async delete(obj: Partial<IFavoriteDeveloper>): Promise<boolean> {
+        const developers = await this.developerStorageService.get(obj);
+        if (developers.length === 0) {
+            return false;
+        }
+        const {username, channelId} = developers[0];
+        const developerRemovingOperationResult = await this.developerStorageService.remove({username, channelId});
         const subscribeRemovingOperationResult = await this.subscribeStorageService.remove({followed: username, channelId});
         return developerRemovingOperationResult && subscribeRemovingOperationResult;
     }
